refactor: 모델 파일명 및 import 명 일치
