+++ src/js/models/LottoStateModel.js
