+++ src/js/models/LottoTicketModel.js
